update v2 driver and remove v3
