--- src/epd7in5_v3/mod.rs
@@ -1,328 +1,0 @@
-//! A simple Driver for the Waveshare 7.5" E-Ink Display (V3) via SPI
-//!
-//! # References
-//!
-//! - [Datasheet](https://www.waveshare.com/wiki/7.5inch_e-Paper_HAT)
-//! - [Waveshare C driver](https://github.com/waveshare/e-Paper/blob/702def0/RaspberryPi%26JetsonNano/c/lib/e-Paper/EPD_7in5_V2.c)
-//! - [Waveshare Python driver](https://github.com/waveshare/e-Paper/blob/702def0/RaspberryPi%26JetsonNano/python/lib/waveshare_epd/epd7in5_V2.py)
-//!
-
-use embedded_hal::{
-    delay::DelayUs,
-    digital::{InputPin, OutputPin},
-    spi::SpiDevice,
-};
-
-use crate::color::TriColor;
-use crate::interface::DisplayInterface;
-use crate::traits::{
-    InternalWiAdditions, RefreshLut, WaveshareDisplay, WaveshareThreeColorDisplay,
-};
-
-pub(crate) mod command;
-use self::command::Command;
-use crate::buffer_len;
-
-/// Full size buffer for use with the 7in5 v3 EPD
-#[cfg(feature = "graphics")]
-pub type Display7in5 = crate::graphics::Display<
-    WIDTH,
-    HEIGHT,
-    false,
-    { buffer_len(WIDTH as usize, HEIGHT as usize * 2) },
-    TriColor,
->;
-
-/// Width of the display
-pub const WIDTH: u32 = 800;
-/// Height of the display
-pub const HEIGHT: u32 = 480;
-/// Default Background Color
-//pub const DEFAULT_BACKGROUND_COLOR: Color = Color::White;
-
-pub const DEFAULT_BACKGROUND_COLOR: TriColor = TriColor::White;
-
-/// Number of bits for b/w buffer and same for chromatic buffer
-const NUM_DISPLAY_BITS: u32 = WIDTH * HEIGHT / 8;
-const IS_BUSY_LOW: bool = true;
-const SINGLE_BYTE_WRITE: bool = false;
-
-/// Epd7in5 (V3) driver
-///
-pub struct Epd7in5<SPI, BUSY, DC, RST, DELAY> {
-    /// Connection Interface
-    interface: DisplayInterface<SPI, BUSY, DC, RST, DELAY, SINGLE_BYTE_WRITE>,
-    /// Background Color
-    color: TriColor,
-}
-
-impl<SPI, BUSY, DC, RST, DELAY> InternalWiAdditions<SPI, BUSY, DC, RST, DELAY>
-    for Epd7in5<SPI, BUSY, DC, RST, DELAY>
-where
-    SPI: SpiDevice,
-    BUSY: InputPin,
-    DC: OutputPin,
-    RST: OutputPin,
-    DELAY: DelayUs,
-{
-    fn init(&mut self, spi: &mut SPI, delay: &mut DELAY) -> Result<(), SPI::Error> {
-        // Reset the device
-        self.interface.reset(delay, 200_000, 4_000);
-
-        // V2 procedure as described here:
-        // https://github.com/waveshare/e-Paper/blob/master/RaspberryPi%26JetsonNano/python/lib/waveshare_epd/epd7in5bc_V2.py
-        // and as per specs:
-        // https://www.waveshare.com/w/upload/6/60/7.5inch_e-Paper_V2_Specification.pdf
-
-        //self.cmd_with_data(spi, Command::BoosterSoftStart, &[0x17, 0x17, 0x27, 0x17])?;
-        self.cmd_with_data(spi, Command::PowerSetting, &[0x07, 0x07, 0x3F, 0x3F])?;
-        self.command(spi, Command::PowerOn)?;
-        self.wait_until_idle(spi, delay)?;
-        self.cmd_with_data(spi, Command::PanelSetting, &[0x0F])?;
-        //self.cmd_with_data(spi, Command::PllControl, &[0x06])?;
-        self.cmd_with_data(spi, Command::TconResolution, &[0x03, 0x20, 0x01, 0xE0])?;
-        self.cmd_with_data(spi, Command::DualSpi, &[0x00])?;
-        self.cmd_with_data(spi, Command::VcomAndDataIntervalSetting, &[0x11, 0x07])?;
-        self.cmd_with_data(spi, Command::TconSetting, &[0x22])?;
-        self.cmd_with_data(spi, Command::SpiFlashControl, &[0x00, 0x00, 0x00, 0x00])?;
-        self.wait_until_idle(spi, delay)?;
-        Ok(())
-    }
-}
-
-impl<SPI, BUSY, DC, RST, DELAY> WaveshareThreeColorDisplay<SPI, BUSY, DC, RST, DELAY>
-    for Epd7in5<SPI, BUSY, DC, RST, DELAY>
-where
-    SPI: SpiDevice,
-    BUSY: InputPin,
-    DC: OutputPin,
-    RST: OutputPin,
-    DELAY: DelayUs,
-{
-    fn update_color_frame(
-        &mut self,
-        spi: &mut SPI,
-        delay: &mut DELAY,
-        black: &[u8],
-        chromatic: &[u8],
-    ) -> Result<(), SPI::Error> {
-        self.update_achromatic_frame(spi, delay, black)?;
-        self.update_chromatic_frame(spi, delay, chromatic)
-    }
-
-    /// Update only the black/white data of the display.
-    ///
-    /// Finish by calling `update_chromatic_frame`.
-    fn update_achromatic_frame(
-        &mut self,
-        spi: &mut SPI,
-        _delay: &mut DELAY,
-        black: &[u8],
-    ) -> Result<(), SPI::Error> {
-        self.interface.cmd(spi, Command::DataStartTransmission1)?;
-        self.interface.data(spi, black)?;
-        Ok(())
-    }
-
-    /// Update only chromatic data of the display.
-    ///
-    /// This data takes precedence over the black/white data.
-    fn update_chromatic_frame(
-        &mut self,
-        spi: &mut SPI,
-        delay: &mut DELAY,
-        chromatic: &[u8],
-    ) -> Result<(), SPI::Error> {
-        self.interface.cmd(spi, Command::DataStartTransmission2)?;
-        self.interface.data(spi, chromatic)?;
-
-        self.wait_until_idle(spi, delay)?;
-        Ok(())
-    }
-}
-
-impl<SPI, BUSY, DC, RST, DELAY> WaveshareDisplay<SPI, BUSY, DC, RST, DELAY>
-    for Epd7in5<SPI, BUSY, DC, RST, DELAY>
-where
-    SPI: SpiDevice,
-    BUSY: InputPin,
-    DC: OutputPin,
-    RST: OutputPin,
-    DELAY: DelayUs,
-{
-    type DisplayColor = TriColor;
-    fn new(
-        spi: &mut SPI,
-        busy: BUSY,
-        dc: DC,
-        rst: RST,
-        delay: &mut DELAY,
-        delay_us: Option<u32>,
-    ) -> Result<Self, SPI::Error> {
-        let interface = DisplayInterface::new(busy, dc, rst, delay_us);
-        let color = DEFAULT_BACKGROUND_COLOR;
-
-        let mut epd = Epd7in5 { interface, color };
-
-        epd.init(spi, delay)?;
-
-        Ok(epd)
-    }
-
-    fn wake_up(&mut self, spi: &mut SPI, delay: &mut DELAY) -> Result<(), SPI::Error> {
-        self.init(spi, delay)
-    }
-
-    fn sleep(&mut self, spi: &mut SPI, delay: &mut DELAY) -> Result<(), SPI::Error> {
-        self.wait_until_idle(spi, delay)?;
-        self.command(spi, Command::PowerOff)?;
-        self.wait_until_idle(spi, delay)?;
-        self.cmd_with_data(spi, Command::DeepSleep, &[0xA5])?;
-        Ok(())
-    }
-
-    fn update_frame(
-        &mut self,
-        spi: &mut SPI,
-        buffer: &[u8],
-        delay: &mut DELAY,
-    ) -> Result<(), SPI::Error> {
-        self.wait_until_idle(spi, delay)?;
-        self.cmd_with_data(spi, Command::DataStartTransmission2, buffer)?;
-
-        self.interface.data(spi, buffer)?;
-
-        // Clear the chromatic layer
-        let color = self.color.get_byte_value();
-
-        self.interface.cmd(spi, Command::DataStartTransmission2)?;
-        self.interface.data_x_times(spi, color, NUM_DISPLAY_BITS)?;
-
-        self.wait_until_idle(spi, delay)?;
-
-        Ok(())
-    }
-
-    fn update_partial_frame(
-        &mut self,
-        _spi: &mut SPI,
-        _delay: &mut DELAY,
-        _buffer: &[u8],
-        _x: u32,
-        _y: u32,
-        _width: u32,
-        _height: u32,
-    ) -> Result<(), SPI::Error> {
-        unimplemented!();
-    }
-
-    fn display_frame(&mut self, spi: &mut SPI, delay: &mut DELAY) -> Result<(), SPI::Error> {
-        self.wait_until_idle(spi, delay)?;
-        self.command(spi, Command::DisplayRefresh)?;
-        Ok(())
-    }
-
-    fn update_and_display_frame(
-        &mut self,
-        spi: &mut SPI,
-        buffer: &[u8],
-        delay: &mut DELAY,
-    ) -> Result<(), SPI::Error> {
-        self.update_frame(spi, buffer, delay)?;
-        self.command(spi, Command::DisplayRefresh)?;
-        Ok(())
-    }
-
-    fn clear_frame(&mut self, spi: &mut SPI, delay: &mut DELAY) -> Result<(), SPI::Error> {
-        self.wait_until_idle(spi, delay)?;
-        self.send_resolution(spi)?;
-
-        self.command(spi, Command::DataStartTransmission1)?;
-        self.interface.data_x_times(spi, 0x00, WIDTH * HEIGHT / 8)?;
-
-        self.command(spi, Command::DataStartTransmission2)?;
-        self.interface.data_x_times(spi, 0x00, WIDTH * HEIGHT / 8)?;
-
-        self.command(spi, Command::DisplayRefresh)?;
-        Ok(())
-    }
-
-    fn set_background_color(&mut self, color: TriColor) {
-        self.color = color;
-    }
-
-    fn background_color(&self) -> &TriColor {
-        &self.color
-    }
-
-    fn width(&self) -> u32 {
-        WIDTH
-    }
-
-    fn height(&self) -> u32 {
-        HEIGHT
-    }
-
-    fn set_lut(
-        &mut self,
-        _spi: &mut SPI,
-        _delay: &mut DELAY,
-        _refresh_rate: Option<RefreshLut>,
-    ) -> Result<(), SPI::Error> {
-        unimplemented!();
-    }
-
-    fn wait_until_idle(&mut self, spi: &mut SPI, delay: &mut DELAY) -> Result<(), SPI::Error> {
-        self.interface
-            .wait_until_idle_with_cmd(spi, delay, IS_BUSY_LOW, Command::GetStatus)
-    }
-}
-
-impl<SPI, BUSY, DC, RST, DELAY> Epd7in5<SPI, BUSY, DC, RST, DELAY>
-where
-    SPI: SpiDevice,
-    BUSY: InputPin,
-    DC: OutputPin,
-    RST: OutputPin,
-    DELAY: DelayUs,
-{
-    fn command(&mut self, spi: &mut SPI, command: Command) -> Result<(), SPI::Error> {
-        self.interface.cmd(spi, command)
-    }
-
-    fn send_data(&mut self, spi: &mut SPI, data: &[u8]) -> Result<(), SPI::Error> {
-        self.interface.data(spi, data)
-    }
-
-    fn cmd_with_data(
-        &mut self,
-        spi: &mut SPI,
-        command: Command,
-        data: &[u8],
-    ) -> Result<(), SPI::Error> {
-        self.interface.cmd_with_data(spi, command, data)
-    }
-
-    fn send_resolution(&mut self, spi: &mut SPI) -> Result<(), SPI::Error> {
-        let w = self.width();
-        let h = self.height();
-
-        self.command(spi, Command::TconResolution)?;
-        self.send_data(spi, &[(w >> 8) as u8])?;
-        self.send_data(spi, &[w as u8])?;
-        self.send_data(spi, &[(h >> 8) as u8])?;
-        self.send_data(spi, &[h as u8])
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn epd_size() {
-        assert_eq!(WIDTH, 800);
-        assert_eq!(HEIGHT, 480);
-        assert_eq!(DEFAULT_BACKGROUND_COLOR, TriColor::White);
-    }
-}
