+++ src/epd7in5b_v2/command.rs
@@ -70,17 +70,25 @@ pub(crate) enum Command {
     DualSpi = 0x15,
 
     /// This command builds the VCOM Look-Up Table (LUTC).
-    LutC = 0x20,
+    LutForVcom = 0x20,
     /// This command builds the Black Look-Up Table (LUTB).
-    LutWW = 0x21,
+    LutBlack = 0x21,
     /// This command builds the White Look-Up Table (LUTW).
-    LutBW = 0x22,
+    LutWhite = 0x22,
     /// This command builds the Gray1 Look-Up Table (LUTG1).
-    LutWB = 0x23,
+    LutGray1 = 0x23,
     /// This command builds the Gray2 Look-Up Table (LUTG2).
-    LutBB = 0x24,
+    LutGray2 = 0x24,
     /// This command builds the Red0 Look-Up Table (LUTR0).
-    LutBD = 0x25,
+    LutRed0 = 0x25,
+    /// This command builds the Red1 Look-Up Table (LUTR1).
+    LutRed1 = 0x26,
+    /// This command builds the Red2 Look-Up Table (LUTR2).
+    LutRed2 = 0x27,
+    /// This command builds the Red3 Look-Up Table (LUTR3).
+    LutRed3 = 0x28,
+    /// This command builds the XON Look-Up Table (LUTXON).
+    LutXon = 0x29,
 
     /// LUT Option
     LutOpt = 0x2A,
